+++ src/views/filters/elements/input.blade.php
@@ -4,15 +4,12 @@
     </label>
     <div>
         <input type="{{ $filter['type'] }}" 
-        @if($data)
-            value="{{ $data }}" 
-        @else
-            placeholder="Wrong Data Passed!" 
-        @endif
+            value="{{ $data ?? '' }}" 
+            placeholder="{{ $filter['placeholder'] }}" 
         name="{{ $filter['column_name'] }}" id="{{ $filter['id'] }}" 
         @if($filter['type'] == 'date') 
             data-date-format="YYYY-MM-DD" 
-        @endif 
+        @endif
         class="form-control data-filters" 
         filter_through_join="{{ $filter['filter_through_join'] }}"
         join_table="{{ $filter['join_table'] }}"